+++ src/Table/Name.php
@@ -20,6 +20,16 @@ final class Name
     }
 
     /**
+     * @psalm-pure
+     *
+     * @param non-empty-string $value
+     */
+    public static function of(string $value): self
+    {
+        return new self($value);
+    }
+
+    /**
      * @return non-empty-string
      */
     public function toString(): string